fix #2934
