+++ ioctl/cmd/account/accountethaddr.go
@@ -10,13 +10,13 @@ import (
 	"fmt"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/spf13/cobra"
-
 	"github.com/iotexproject/iotex-address/address"
+	"github.com/spf13/cobra"
 
 	"github.com/iotexproject/iotex-core/ioctl/config"
 	"github.com/iotexproject/iotex-core/ioctl/output"
 	"github.com/iotexproject/iotex-core/ioctl/util"
+	"github.com/iotexproject/iotex-core/pkg/util/addrutil"
 )
 
 // Multi-language support
@@ -62,7 +62,7 @@ func accountEthaddr(arg string) error {
 		}
 		ioAddr = ioAddress.String()
 	} else {
-		ethAddress, err = util.IoAddrToEvmAddr(ioAddr)
+		ethAddress, err = addrutil.IoAddrToEvmAddr(ioAddr)
 		if err != nil {
 			return output.NewError(output.AddressError, "", err)
 		}
