+++ ioctl/util/util.go
@@ -18,16 +18,12 @@ import (
 	"strings"
 	"syscall"
 
-	"google.golang.org/grpc/metadata"
-
-	"github.com/ethereum/go-ethereum/common"
 	"github.com/spf13/cobra"
 	"go.uber.org/zap"
 	"golang.org/x/crypto/ssh/terminal"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/credentials"
-
-	"github.com/iotexproject/iotex-address/address"
+	"google.golang.org/grpc/metadata"
 
 	"github.com/iotexproject/iotex-core/ioctl/config"
 	"github.com/iotexproject/iotex-core/ioctl/output"
@@ -104,15 +100,6 @@ func RauToString(amount *big.Int, numDecimals int) string {
 		return amountInt.String() + "." + decString
 	}
 	return amountInt.String()
-}
-
-// IoAddrToEvmAddr converts IoTeX address into evm address
-func IoAddrToEvmAddr(ioAddr string) (common.Address, error) {
-	address, err := address.FromString(ioAddr)
-	if err != nil {
-		return common.Address{}, output.NewError(output.ConvertError, "", err)
-	}
-	return common.BytesToAddress(address.Bytes()), nil
 }
 
 // StringToIOTX converts Rau string to Iotx string
